fix: Rename field mixin with displayed column.
